Section 3.1: Outlining the first test case.
Although in the real world, this would not be it's own commit as it
is not atomic (as the tests fail at this point) I thought for
educational purposes following the video tutorials it could be good
to have each episode as it's own commit. This is so incremental
changes can be followed.